requirements.txt
transformers
torch
pymongo
python-dotenv # For managing configuration securely
flask # For web interface
huggingface-hub # For authentication
accelerate # Required for device_map='auto'
